refactoring code structure for better server/client components interaction
