+++ components/shared/Icon/IconAvailable.tsx
