+++ components/shared/Icon/IconSelected.tsx
