+++ components/shared/ThemeToggle/styles.ts
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+
+export const StyledThemeToggle = styled.button`
+  position: fixed;
+  top: 1rem;
+  right: 1rem;
+  padding: 0.5rem 1rem;
+  border-radius: 20px;
+  background: var(--accent, #fd6d8e);
+  color: var(--background, #ffffff);
+  border: none;
+  cursor: pointer;
+  font-size: 0.9rem;
+  box-shadow: 0 2px 5px -4px currentColor;
+`;
