+++ hooks/useTheme.ts
@@ -1,0 +1,23 @@
+'use client';
+
+import { createContext, useContext } from 'react';
+
+export type Theme = 'light' | 'dark';
+
+interface ThemeContextType {
+  theme: Theme;
+  setTheme: (theme: Theme) => Promise<void>;
+}
+
+export const ThemeContext = createContext<ThemeContextType>({
+  theme: 'dark',
+  setTheme: async () => {},
+});
+
+export function useTheme() {
+  const context = useContext(ThemeContext);
+  if (!context) {
+    throw new Error('useTheme must be used within a ThemeProvider');
+  }
+  return context;
+}
